added verification of final apk signature
It verifies if a classes.dex file (which is in every apk) was signed using
proper key. In debug configuration, that is Android default debug key,
in release configuration, that is our own testing key.

Verifying single file is enough for purpose of these tests, but we might
verify all files in the apk as well.