'clean' task is not tested on Windows (hack)
So, this is officially driving me nuts.

Switching the zipalign implementation to Gradle's Exec helped against file
handle leaks, but only partially. Tests are still failing, but less often.
This is unacceptable, so on Windows, I completely switched off testing the
'clean' task, which was the one failing because of file handle leaks.