automatically enable Proguard when compiling Scala project
Trying to dex unproguarded JAR of Scala project fails, so we automatically
enable Proguard in such condition. Someone might think that we are doing
magic here, enabling something that is normally disabled, but since it is
really _necessary_ to build the project successfully, I think we should do it.
And it gets logged on INFO level, at least.