added an example (and a test) of another way of configuring release build
The preferred way of configuring debug/release builds with 'configureDebug'
and 'configureRelease' tasks is kind of error-prone and not much pretty.
I personally would like simple 'build' task to produce debug artifacts and
another 'release' task to produce release artifacts. This new test shows
exactly how to do that. It is a little convoluted, because Gradle always
executes configuration closures of all defined tasks, not only those planned
for execution, but it works and shows few important aspects of Gradle.