adjusted the way integration tests are launched so that they are still running
Gradle 1.0-m3 or m4 changed the set of classes that the buildscript has access
to. Even if it has some classes on the classpath, the buildscript can only
access a few selected packages from Gradle and relatives. But our integration
tests currently depend on having Gradle Android Plugin available to the tests
buildscripts. This commit is a hack to allow this.

Since GradleLauncher was deprecated in Gradle 1.0-m3, it isn't so bad to hack
this up as we should switch to Tooling API anyway. However, we can't do it yet,
because the Tooling API currently doesn't let you specify the buildscript file
name and it assumes that it will always be called 'build.gradle'. On the other
hand, we heavily use custom buildscript file names, e.g. the 'hello' project
has 3 buildscripts for different tests. This will be supported in the future,
as Adam Murdoch told me on the Gradle Users mailing list.