Include changes that will allow us to upload to Sonatype's OSS
This includes having the signing plugin, as well as adding the snapshot
and staging repositories to uploadArchives.

Note that a gradle.properties file should be created to store the GPG
and Sonatype repository information.  My gradle.properties was not
committed as it has sensitive information.  The following items should
be present in gradle.properties:

    signing.keyId=
    signing.password=
    signing.secretKeyRingFile=
    sonatypeUsername=
    sonatypePassword=

Signed-off-by: Ealden Esto E. Escanan <ealden@gmail.com>