Initial work on more flexible and reusable tasks.
All tasks declare their inputs and outputs, to benefit from Gradle's
up-to-date checking. The PluginConvention class has been replaced with
a new-style PluginExtension.

The plugin supports library and test projects natively, and manages
the dependencies between them. It also has tasks for Emma code
coverage and Android Lint.

The plugin has been tested with r17 of the Android SDK tools and
Gradle 1.0. Compatibility with earlier revisions of the SDK was not an
objective.

Signed-off-by: Marcus Better <marcus@better.se>